+++ clikt/src/main/kotlin/com/github/ajalt/clikt/parameters/options/ValueWithDefault.kt
@@ -1,0 +1,7 @@
+package com.github.ajalt.clikt.parameters.options
+
+/** A container for a value that can have a default value and can be manually set */
+data class ValueWithDefault<out T>(val explicit: T?, val default: T) {
+    val value: T get() = explicit ?: default
+}
+